[native] Introduce `inlineSidebar` animation when opening message tooltip to avoid visual issues.
Summary:
Fix visual issues when opening message tooltip. For the animation, it requires rendering the second `InlineSidebar` component on the top, so it could diffuse the message.

Before:
{F136210}
After:
{F136217}

Test Plan: Tested the animation on both platforms for multimedia and text messages. Confirmed if "go to sidebar" animation is not affected.

Reviewers: tomek, atul, abosh, ashoat

Reviewed By: tomek, atul, abosh, ashoat

Subscribers: ashoat, adrian

Differential Revision: https://phab.comm.dev/D4825