[native] Redesign `InlineSidebar` for reactions feature
Summary:
Introduce redesigned `InlineSidebar` component for native. It is now able to keep both sidebar and reactions.
Now it receives reactions as an array of strings - so it would be easy to test visually by providing an array of emojis. Probably, in the future we will keep reactions in some more complex data structure, so the types should be adjusted then.

---

{F136176}
{F136177}
{F136202}
{F136204}
GIF:
{F136208}

Test Plan:
Tested the component look on both platforms in both color themes. Also provided smoji and tested displaying reactions and/or sidebar.
The visual issues when opening message tooltip will be fixed in next diff.

Reviewers: tomek, atul, abosh, ashoat

Reviewed By: tomek, atul, abosh, ashoat

Subscribers: ashoat, adrian

Differential Revision: https://phab.comm.dev/D4824