[lib] Make `useInlineSidebarText` hook parameter optional
Summary:
When we add reactions to `InlineSidebar` component we want to support the case, when there are reactions and there is no existing thread from the message.
In that case, we don't have `ThreadInfo`. The idea is to make the hook return empty strings when no there is `threadInfo` provided. It will help avoid conditional execution of the hook in that case.

Test Plan: Current behavior remain unchaged. After introducing `InlineSidebar` redesign, confirmed that the hook returns empty result strings when there is no thread provided.

Reviewers: tomek, atul, abosh

Reviewed By: tomek, abosh

Subscribers: ashoat, adrian

Differential Revision: https://phab.comm.dev/D4823