[native] add Rust gRPC client to XCode project
Summary:
This change adds the static library for the gRPC client to XCode. Now the functions can be referenced from other C++ files in the project. I had to add a new dependency, cargo lipo, to build a universal static library that works on both x86 and arm since the simulator doesn't use arm for some reason...

Depends on D4951

Test Plan: built the app for physical iOS devices and for the simulator

Reviewers: tomek, atul, marcin, jon

Reviewed By: tomek, atul, marcin, jon

Subscribers: inka, jon, ashoat, abosh

Differential Revision: https://phab.comm.dev/D4953