[landing] SIWE component with RainbowKit that connects & signs a message
Summary:
The goal: to sign in with ethereum! to do that we need a signed message from the wallet to pass to the identity service (which is currently native code, not in the browser).

This diff adds a top level routing page that includes RainbowKit's Connect Wallet button, as well as a function that will pass the signed message back from the webview to native where the identity service code lives.

More context for why we're making these changes in landing, why there are so many package changes, and what other options have been considered can be found in [[ https://linear.app/comm/issue/ENG-1212/sign-in-with-ethereum | linear, issue 1212 ]].

Test Plan: run `/landing` and open `/commlanding/siwe` on a simulator webview to get a message over the bridge

Reviewers: abosh, atul, ashoat, varun

Reviewed By: ashoat, varun

Subscribers: jacek, varun, ashoat, tomek, adrian, atul, abosh

Differential Revision: https://phab.comm.dev/D4510