[services] Backup - Rust lib - Add process
Summary:
Depends on D4867

Adding function to schedule data in the thread-safe queue in rust to be called from the c++.

The goal here is to:
- spawn a thread in rust that will listen on a thread-safe queue (`initialize`)
- **schedule some data on this queue** (`process`)
- gracefully exit the thread and release all resources (`terminate`)

Later on, in the spawned thread, we will connect to the blob service through the gRPC, but first things first.

I'm probably going to remove all logging before landing, I'm leaving it as is for now.

Test Plan: test plan in D4870

Reviewers: tomek, varun, max

Reviewed By: tomek

Subscribers: ashoat, tomek, atul, abosh

Differential Revision: https://phab.comm.dev/D4869