[native] Upload video thumbnails to `keyserver` and persist in serverDB
Summary:
This diff uploads the local thumbnail generated on the `native` client to the `keyserver` where the contents will be persisted in the `uploads` table of the `serverDB`.

NOTE: This diff can be landed as-is. The entry's `container` column will be `NULL` since there's still plenty of work we need to do in `multimediaMessageCreationResponder(...)` before we're ready to call `assignMedia(...)` (or the equivalent) to set the `container` value. However, it's still fine to add these entries to the `uploads` table because they'll be swept up by the `deleteUnassignedUploads` (`keyserver/src/deleters`) cron job that runs every day at 3:30AM PT (`keyserver/src/cron/cron.js`).

Test Plan:
1. Send a video message on `native` (iOS Simulator)
2. Open `uploads` table in the serverDB (MariaDB)
3. Check to see whether there are two rows for the sent media message. One with the contents of the video, and another with the thumbnail image:
{image https://blob.sh/250ade.png}

Reviewers: abosh, tomek, jacek, ashoat

Reviewed By: abosh, tomek

Differential Revision: https://phab.comm.dev/D4979