[native] Update local video messages with realized `thumbnailURI` upon successful upload
Summary:
Modify payload for `updateMultimediaMessageMedia` action to include realized `thumbnailURI` upon successful upload when `mediaType === "video"`.

We need to update the `thumbnailURI` because the temporary local thumbnail gets `unlink`ed after a successful upload.

---

Depends on D4979

Test Plan: Used the Redux remote dev tools in Chrome to observe the local media message (and the contents of the local media array within) and make sure that the local `thumbnailURI` got replaced with the "realized" (aka server) `thumbnailURI` as expected (on dispatch of `updateMultimediaMessageMediaActionType`)

Reviewers: abosh, jacek, tomek, ashoat

Reviewed By: abosh, tomek

Differential Revision: https://phab.comm.dev/D4980