Share database worker thread between CommCoreModule and AppDelegate
Summary: This differential is a refactor that enables to share database worker thread between AppDelegate and CommCoreModule. We suspect that the reason for iOS 137 crash is that we access SQLite database from thread upon rescind when some other database operations is being executed on databaseThread of CommCoreModule. Sharing database thread with CommCoreModule and AppDelegate should remedy the crash then. IMPORTANT: this revision is not going to be landed. We will create release build with its commit, push to the master and revert it. Therefore it is acceptable that code here is not 100% clean. We only want it to work to either confirm or refute our theory. If it happens to be a remedy for the crash we will create another one with clean OOP design that will be traditionally landed. Review here should focus on correctness and edge cases that might not have been tested.

Test Plan: Release build created from commit represented by this revision should be free from iOS 137 crashes even though it will fully unlock rescinding functionality.

Reviewers: tomek, atul

Subscribers: