[landing] SIWE component with RainbowKit that connects & signs a message
Summary:
take two of D4510, which was reverted because of breaking css changes. re-pasting the summary here

The goal: to sign in with ethereum! to do that we need a signed message from the wallet to pass to the identity service (which is currently native code, not in the browser).

This diff adds a top level routing page that includes RainbowKit's Connect Wallet button, as well as a function that will pass the signed message back from the webview to native where the identity service code lives.

More context for why we're making these changes in landing, why there are so many package changes, and what other options have been considered can be found in linear, issue 1212.

Test Plan: run /landing and open /commlanding/siwe on a simulator webview to get a message over the bridge

Reviewers: abosh, atul, ashoat, varun

Subscribers: abosh, atul, adrian, tomek, ashoat, varun, jacek

Differential Revision: https://phab.comm.dev/D4997