[lib] Introduce `tMediaMessage[Photo/Video/Media]` to validate requests to `textMessageCreationResponder`
Summary:
We're modifying `textMessageCreationResponder` to be able to accept a list of `MediaMessageServerDBContent`s as one of the arguments.

Before we build the new validator for `textMessageCreationResponder` we introduce these basic types that we will build upon in subsequent diffs.

---

Depends on D4985

Test Plan: Will add tests in next diff

Reviewers: abosh, jacek, tomek, ashoat

Reviewed By: abosh, tomek

Differential Revision: https://phab.comm.dev/D4990