[DRAFT] [services] Blob - check for holder/hash presence in terminate callback
Summary:
Depends on D4942

We want to check if the mandatory fields are present in `terminateCallback`, because we cannot know whether the client will not just terminate the connection. In such a case, we should identify the first error that prevented the reactor from working properly.

Test Plan: Test plan in D4946

Reviewers: tomek, jon

Reviewed By: tomek

Subscribers: ashoat, tomek, atul, abosh

Differential Revision: https://phab.comm.dev/D4943