[native] Select thread name when clicking `EditSettingButton`
Summary:
Related Linear issue [[ https://linear.app/comm/issue/ENG-1730/select-thread-name-when-clicking-edit-button-on-native | here ]].

This diff selects the thread name when clicking the edit button in `ThreadSettingsName`, and also allows the viewer to view the entire thread name on multiple lines instead of only viewing a fixed portion of the thread name on a single line.

Test Plan:
Tested on iOS and works as expected.

Before:

After:

On Android, it looks like neither the description nor thread name are selected when clicking on them. Not sure if this is expected behavior on Android (it existed prior to me making any changes), but this diff doesn't modify anything on Android.

Reviewers: atul, tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D5032