[services] Rust Integration - Backup - c++ - Multiple put clients
Summary:
Depends on D5035

In order to be able to handle multiple connections at the same time, we have to store a collection of clients in the rust state instead of having just one.

Here, I'm making the c++ calls to rust for the put client valid.

Test Plan:
```
cd services
yarn run-integration-tests backup
```

Reviewers: jon, tomek, max

Reviewed By: tomek

Subscribers: ashoat, tomek, atul, abosh

Differential Revision: https://phab.comm.dev/D5036