[services] Rust Integration - Backup - Rust - Multiple get clients
Summary:
Depends on D5036

In order to be able to handle multiple connections at the same time, we have to store a collection of clients in the rust state instead of having just one.

Here, I'm adding the collection to the get blob client.

Test Plan:
The backup service will not build successfully because of invalid calls from the c++ but I decided to divide the code so the diffs are not too big. This is fixed in the next diff in the stack - D5038

This will work:
```
cd services/backup/blob_client
cargo check
```

Reviewers: jon, tomek, max, varun

Reviewed By: tomek

Subscribers: ashoat, tomek, atul, abosh

Differential Revision: https://phab.comm.dev/D5037