[web] Redesign `InlineSidebar` with reactions support
Summary:
Introduce redesigned InlineSidebar component.
Currently it receives reactios as an array of strings to simplify testing with emojis, but it the future it will problably need to receive some complex rection data type.

{F146977}

Test Plan: Tested text, media and robotext messages in all cases with/without sidebar or/and reactions.

Reviewers: tomek, atul, abosh

Reviewed By: tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D4924