[web] Remove old tooltip components
Summary:
Remove no longer used code related to tooltip.
At this point, only the new tooltip should be displayed.

The new component on the video:
https://linear.app/comm/issue/ENG-1575#comment-d3fe1db0

Test Plan: Tested the new tooltip on text, media and robotext messages with different sizes a location. Tested if all the actions work as before and the action list remains unchanged. The new tooltip should not cover inline sidebar if it exists.

Reviewers: tomek, atul, abosh

Reviewed By: tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D4923