[services] Rust Integration - Backup - Rust - Fix wrong order of logs in the tests
Summary:
Depends on D5071

Since some logs are pulled later and some sooner, we have to identify them by IDs instead of the order in which they come from the server. For example, if you have the logs of size 3MB, 3MB, 5KB, or 3MB, first you'll probably receive this of 5KB and only then the rest.

Test Plan: integration tests should pass for backup.

Reviewers: jon, max, tomek, varun

Reviewed By: tomek

Subscribers: ashoat, tomek, atul, abosh

Differential Revision: https://phab.comm.dev/D5072