[native] remove dependency on identity service package for gRPC client
Summary:
The location of the protos directory is still in flux. That means that it's difficult to take a dependency on the Identity service crate, even though it would be ideal to not repeat OPAQUE config code on client and server. This issue tracks restoring the dependency on the Identity crate: https://linear.app/comm/issue/ENG-1746/remove-opaquers-from-grpc-client

Added ashoat since there are new dependencies (these are all in the Identity service as well)

Test Plan: `cargo build`

Reviewers: tomek, atul, ashoat, jon

Reviewed By: tomek, atul, ashoat

Subscribers: abosh, jon, marcin, max, karol, inka

Differential Revision: https://phab.comm.dev/D5075