[lib] Lowercase all reserved usernames
Summary:
Somebody was able to register with the username "comm". Looked into it and found that we were checking lowercased usernames against the list, but some of the entries in the list were not lowercase.

Linear issue: [ENG-1756](https://linear.app/comm/issue/ENG-1756/make-sure-reserved-username-list-is-all-lowercase)

Test Plan: I used vim to search for the RegExp `[A-Z]` in the list and found no results

Reviewers: atul, tomek

Reviewed By: atul

Subscribers: abosh

Differential Revision: https://phab.comm.dev/D5080