[web] Added possibility to change input type and icon position in EnumSettingsOption
Summary:
Depends on D5110
Context: [[ https://linear.app/comm/issue/ENG-1767/extend-enumsettingoption-with-a-checkbox | here ]]

Added possibility to change input type of `EnumSettingsOption` from "radio" to "checkbox" and added `iconPosition` param to set vertical position of the left icon.

Additionally, setting cursor icon to pointer, when we hover on it.

Now, according to Figma design, we could use one component to create both radio and checkbox options.
{F167124}

Test Plan:
Currently, this improvement hasn't been used yet. It will be used in next diff.

To change `EnumSettingsOption` input type to checkbox, set `inputType` param as a "checkbox".

To change a vertical position of the left icon, set `iconPosition` as one of the {"top", "center", "bottom"} (by default is set to "center")

Reviewers: tomek, atul, ashoat

Reviewed By: tomek, atul

Subscribers: ashoat, tomek, atul, abosh

Differential Revision: https://phab.comm.dev/D5115