[keyserver] Modify `legacyMultimediaMessageCreationResponder` to handle new payload
Summary:
This is a temporary change on the `keyserver` side that keeps things working as they are when client hits `create_multimedia_message` with the new payload.

We're basically translating `mediaMessageContents` back to `mediaIDs` by pulling out the `uploadID` property.

This ensures that the client continues working as normal with the new payload while we're still working on implementing the "new" `multimediaMessageCreationResponder`.

---

Depends on D5061

Test Plan:
Able to send image(s) on `web` and `native` as expected.

Set breakpoints at various points throughout the media message flow and observed that values were as expected at each point.

Reviewers: abosh, tomek, marcin

Reviewed By: tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D5062