[services][identity] add public key to pake credential request
Summary: if a user doesn't have a keyserver, we need each peer to be able to send messages to all of that user's devices simultaneously. in order to do that, we need the public keys for all of their devices. so, when a user logs in with a password on a device, they should provide the public key for that device.

Test Plan: proto change, built the identity service and client

Reviewers: tomek, jon

Reviewed By: tomek

Subscribers: ashoat, atul, karol, abosh

Differential Revision: https://phab.comm.dev/D5130