[services] Tunnelbroker - Add device ping interval to Constants
Summary:
This diff introduces a ping interval to check if the device is online.
The default "beacon" value for pinging is `3000` ms in our keyserver, which seems reasonable.
More context can be found in the Linear task.

Related Linear task: [[ https://linear.app/comm/issue/ENG-1766/make-sure-the-device-is-offline-to-send-a-notification-to-wake-it-up | ENG-1766 ]]

Test Plan: No test, just a constant value.

Reviewers: jon, varun, tomek, karol, ashoat

Reviewed By: tomek, ashoat

Subscribers: ashoat, tomek, atul, karol, abosh

Differential Revision: https://phab.comm.dev/D5151