[keyserver] Introduce `assignMessageContainerToMedia(...)` in `upload-updaters`
Summary:
Pretty much `assignMedia(...)` but accepts `mediaMessageContents: $ReadOnlyArray<MediaMessageServerDBContent>` instead of `mediaIDs: $ReadOnlyArray<string>`.

The intention is to use `assignMessageContainerToMedia(...)` in the "new"/non-legacy `multimediaMessageCreationResponder(...)`

---

Depends on D5086

Test Plan:
Later in the stack this function is "consumed" from `multimediaMessageCreationResponder(...)`. I set some breakpoints and confirmed that values were as expected. I also checked the `uploads` table in my local MySQL DB and found that the `container` column was assigned as expected:

{F164479}

Reviewers: abosh, tomek, marcin

Reviewed By: tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D5087