[services] Tunnelbroker - Update `SessionItem` database entity with the `IsOnline` field
Summary:
This diff introduces adding of `IsOnline` field to the DynamoDB database in the `DeviceSessionItem` entity.
To send notifications if the device is offline we need to distinguish the device state. To do this we will ping the device and change the device status (online/offline) in this database in this field.

This diff is a part of the stack.
More context is available in the Linear task description.

Related Linear task: [[ https://linear.app/comm/issue/ENG-1766/make-sure-the-device-is-offline-to-send-a-notification-to-wake-it-up | ENG-1766 ]]

Test Plan:
1. Successfully built.
2. Run `yarn run-unit-tests` and `yarn run-integration-tests`.

Reviewers: jon, varun, tomek, karol

Reviewed By: jon, tomek

Subscribers: ashoat, tomek, atul, karol, abosh

Differential Revision: https://phab.comm.dev/D5164