[services] Tunnelbroker - Add `updateSessionItemIsOnline` database method
Summary:
This diff introduces adding of `updateSessionItemIsOnline` database method.

To send notifications if the device is offline we need to distinguish the device state. To do this we will ping the device and change the device status (online/offline) in the `IsOnline` field (D5164) using the introduced method.

More context is available in the Linear task description.

Related Linear task: [[ https://linear.app/comm/issue/ENG-1766/make-sure-the-device-is-offline-to-send-a-notification-to-wake-it-up | ENG-1766 ]]

Test Plan:
1. Successfully built.
2. Passed a unit test in D5166.

Reviewers: jon, tomek, karol

Reviewed By: tomek

Subscribers: ashoat, tomek, atul, karol, abosh

Differential Revision: https://phab.comm.dev/D5165