[services] Tunnelbroker - Add test for the `updateSessionItemIsOnline` database method
Summary:
This diff introduces a test for the `updateSessionItemIsOnline` method from D5165.

In this test, we will create a `SessionItem` entity and push it to the DynamoDB table. The default value for the `IsOnline` (D5164) field is `false`. We are checking the default value and updating it to `true` using the `updateSessionItemIsOnline` method and checking the value to become `true` after.

Related Linear task: [[ https://linear.app/comm/issue/ENG-1766/make-sure-the-device-is-offline-to-send-a-notification-to-wake-it-up | ENG-1766 ]]

Test Plan:
1. Successfully built.
2. The test was successfully passed.

Reviewers: jon, tomek, karol

Reviewed By: tomek

Subscribers: ashoat, tomek, atul, karol, abosh

Differential Revision: https://phab.comm.dev/D5166