[lib] Persist messages after user account is deleted
Summary:
After a user deletes their account, there is a bug that deletes
their messages from another user's device and removes any trace of
conversation. This diff addresses that by instead persisting the messages,
but replacing the deleted user's account name with anonymous to signify a
deleted account.

Addresses: https://linear.app/comm/issue/ENG-1735/messages-of-a-deleted-account-dont-show-up-in-the-app

Test Plan:
Simulated a conversation between two individuals and observed
that, as expected, the deleted user is renamed to anonymous and their
messages persist. In a group chat, the same behavior is expected, where
the user is removed from the group but their messages persist and it is
authored by anonymous.

Video Demo:
{F176142}

Reviewers: atul, varun, ashoat

Reviewed By: ashoat

Subscribers: ashoat, tomek, abosh

Differential Revision: https://phab.comm.dev/D5199