[services] Tunnelbroker - Update Protobuf file with the `ping`
Summary:
This diff introduces adding of `ping` in the `GetResponse` message for the Tunnelbroker API to check if the device is online.
This `ping` message will be sent to the client in a ping interval introduced in D5151. By pinging the client in a unidirectional `Get` stream we will be sure that the client is online and update the state of the client in the database using the method from D5165.
We can use the same approach when changing to the bidirectional stream in a near future in v.0.5.

This is a part of the stack.

Related Linear task: [[ https://linear.app/comm/issue/ENG-1766/make-sure-the-device-is-offline-to-send-a-notification-to-wake-it-up | ENG-1766 ]]

Test Plan: Successfully built the server and the follow-up diffs in the stack.

Reviewers: jon, karol, ashoat

Reviewed By: jon, ashoat

Subscribers: ashoat, tomek, atul, karol, abosh

Differential Revision: https://phab.comm.dev/D5173