[native] Increase size of camera and gallery icons
Summary:
Increased the size of the icons in chat (camera and gallery icons), and added some more padding
between the icons and the chat input to address the issue of accidental clicks. This addresses ENG-1723:
https://linear.app/comm/issue/ENG-1723/take-photo-photo-gallery-icons-should-be-bigger

Test Plan:
Compared the icon sizes on both Android and iOS before and after making these changes, and the
icons are visually larger and easier to click as they are separated a little more. Since there is no official
design for this yet, reviewer feedback can decide if it's visually and functionally better.

Before:
{F172392}

After:
{F172393}

Additionally, I tested the tap radiuses of each component and it's much easier to interact with the respective icons. Similarly, increasing the size of these doesn't cause any undesired behavior with the expando button and typing a message.

Reviewers: atul, abosh, ashoat

Reviewed By: atul, ashoat

Subscribers: ginsu, ashoat, tomek

Differential Revision: https://phab.comm.dev/D5169