[lib] Introduce new `sendMultimediaMessage` and consume in `native`
Summary:
Introduce a new `sendMultimediaMessage` in `lib/actions/message-actions` that hits the `create_multimedia_message` endpoint with the updated payload (`mediaMessageContents` instead of `mediaIDs`).

This endpoint is consumed by `input-state-container` on `native`, but **not** on `web`. We continue to use the `legacySendMultimediaMessage` on `web` to avoid breaking things.

IMPORTANT: While the changes made to `keyserver` in this stack are backwards compatible, this change to `native` depends on `keyserver` being deployed before a release is made. To keep things simple I'll avoid landing this diff until all necessary changes on the `keyserver` side are made and tested. I'll hold back on landing this diff (and a few others) until we decide to "flip the switch." I'll probably just keep this commit in my local environment as I'm working, but figured it might be helpful for reviewers if they wanted to `arc patch` or test any part of the stack in their local dev environment.

(As of this diff, the `keyserver` has everything it needs in `multimediaMessageCreationResponder(...)` to properly associate uploads with media with messages)

---

Depends on D5062

Test Plan:
1. Able to send image(s) on `native` as before without issue
2. Able to send image(s) on `web` as before without issue

Also set breakpoints at key parts of the multimedia message flow and ensured that values appeared as expected.

(I'll do some a bunch of additional testing once things are "working" end-to-end)

Reviewers: abosh, tomek, marcin

Reviewed By: tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D5063