[native] move threadInfo caching logic
Summary:
We no longer use navigation params directly in the inner component and moving caching logic of threadInfo to the outer component makes that clearer.

Depends on D5142

Test Plan: Check that if you change some value in settings (e.g. color) and log out, the settings screen that "blinks" for a moment has an updated value.

Reviewers: tomek, atul, abosh, ashoat

Reviewed By: ashoat

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D5143