[services][identity] change DDB schema for public keys to map
Summary: We need to store a public key for each device in the users table in ddb. Currently, we only store one public key. This diff changes the table schema to include a mapping of device IDs to public keys.

Test Plan: modified `main.rs` to call the `update_users_table` method, and verified that the map was updated appropriately when repeat and new keys were set in the update expression.

Reviewers: tomek, jon, atul

Reviewed By: tomek, jon

Subscribers: ashoat, abosh

Differential Revision: https://phab.comm.dev/D5200