[web] Change See More... hover color to match chat threads hover color
Summary:
Changed the div.thread class's hover color to match the hover color of the
activeThread, since var(--thread-active-bg) is a different shade from
var(--thread-hover-bg).

The difference is shown in the `themes.css` file:

```
  --thread-hover-bg: var(--shades-black-90);
  --thread-active-bg: var(--shades-black-80);
```

Test Plan:
Ran the before and after versions of this change and visually
the colors are the same with this change.

Before:
{F171814}

After:
{F171815}

For a closer look at the change in the thread color:

Before:
{F171864}

After:
{F171865}

Reviewers: atul, varun

Reviewed By: atul, varun

Subscribers: varun, ashoat, tomek, abosh

Differential Revision: https://phab.comm.dev/D5162